requirements.txt
bilibili_api_python==16.2.0
edge_tts==7.0.2
Flask==2.3.2
jieba==0.42.1
python-dotenv==1.2.1
Requests==2.32.5
urllib3==1.26.11
